// src/__generated__/VillagersFilterQuery.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: VillagersFilterQuery
// ====================================================

export interface VillagersFilterQuery_villagersFilter {
  __typename: "VillagersFilterOutput";
  ok: boolean;
  species: string[] | null;
  personalities: string[] | null;
}

export interface VillagersFilterQuery {
  villagersFilter: VillagersFilterQuery_villagersFilter;
}